Add protection to linear multiply and fix memory leak
Previously never freed the weight and bias matrix which caused a memory
leak in the program. This is now fixed.

Also added an assert statement to the multiply function to ensure the
width and _in_feature dimensions are the same (as required by matrix
multiplication). Otherwise the error spat out by blas would be more
cryptic.