Add matrix gradient backpropagation
I wrote layer.h which handles backpropagation. It computes the error
gradient w.r.t. the input, the bias, and the weights. The error gradient
w.r.t. to the input is used to backpropagate the gradient. The other two
gradients are used to update the weight and bias matrixes.

Models.h also now handles all this backpropagation (just for the critic
network though, will have to implement actor network soon). In learn.cc
I wrote a basic script to train the critic network to add three numbers
and multiply them by two (the intention of this is to show how the
backprop works, and to show that it actually does work)

I also updated all the matrixes to not be on the heap since then I don't
have to deal with freeing all of them.