Remove repeated creation of new matricies
Many matricies were created as part of the process of forward, and
backpropagation. This slowed down the program significantly as some
matricies were as large as 256x256 doubles (0.5 MB!). By removing this
unnecesary allocation I sped the program up by about 5x. (I added
benchmarking to learn.cc to measure this speedup)

There were some diminishing returns with some of the other matricies I
optimized, as they were not as large as 256x256. So I need to determine
whether the extra complexity in the code is worth the marginal speedup.
(I'll likely end up finding a neater way to write the code which will
remove the tradeoff as well).