Clean up matrix initialization code
Previously had many large if statements to initialize class matricies on
the first time the corresponding function was called. This helped with
latency as double arrays don't have to be deleted and reallocated
multiple times, but can instead be reused making the program faster.

However a drawback of these if statements is that it made the code
messy, so instead I updated calls to matrix.h functions to setup the
matrix there automatically (so if the matrix is the incorrect size it
will properly update the dimensions, otherwise it will leave it as is)
This made the code much neater and using this I updated critic models.h
to never allocate new double arrays during the forward or backprop
methods.