Remove ununsed parameter from tensor add_ function
One add function took in another tensor and a double, then added the
tensor to the existing tensor, but ignored the double. I updated the
function to no longer request a double as a parameter. This way the
function can still use openblas to do the math (blas doesn't support
adding a constant to each value unfortunately)