Update makefile/dockerfile to support learn.cpp
Dockerfile previously didn't install openblas, however, since
we now need that for learn.cpp I've updated the dockerfile to
install and setup the required environment variables for it. I've
also updated the makefile so it works with the docker container it
didn't link openblas exactly correctly, and some options weren't
supported so I fixed both of those.