Implement basic linear, tensor cuda functionality
I've implemented all the functionality required to construct tensors
and Linear from the device. I've also implemented all the required
functionality to get the linear forward method to work.

There is an issue with the output of the matrix multiplication
function though. I will need to look into it further at a later time
(it will need to be optimized and setup to handle transpose properly
also)