Implement working cuda learning
There is now a cuda kernel set up which is able to learn how to add
three numbers. This currently uses the grad_descent optimizer as
there are issues with the adam optimizer (may be something simple,
haven't had a chance to look at it yet).

There is also an issue when I increase the threads past 32. I'm in
the process of debugging this (currently I suspect something with
using too much memory/registers although I could be wrong)

Also in order to run test.cu you have to make sure the gpu isn't
interuppted (for me that meant I had to disconnect the jetson's
monitor since the display program was interrupting to use the gpu)